Enhancement: Refactor video handling to use Vision namespace, introduce VideoPacket class for better packet management, and update related methods for improved clarity and functionality.

+++ src/Modules/RcCommandAndControl.hpp
@@ -38,7 +38,7 @@ protected:
 
     typedef struct __attribute__((__packed__))
     {
-        uint32_t   sequence_id;
+        uint16_t   sequence_id;
         uint16_t   msg_length;
     } MsgHeader_t;
